codigo limpio
