+++ src/types.d.ts
@@ -96,3 +96,17 @@ interface TableUser {
   role: string;
   active:boolean;
 }
+interface User {
+  id: string; 
+  name: string;
+  lastName: string;
+  email: string;
+  password: string;
+  role: string;
+  active: boolean;
+  deleted: boolean;
+}
+interface UserCardProps {
+  user: User;
+  onDelete: () => Promise<void>;
+}
